requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
WeasyPrint==53.3
